refactor(cuda): dedup CudaMemoryManager by merging Integration facade into canonical
Removes the 1451-LOC Integration/Components/CudaMemoryManager facade that was a
thin wrapper around Memory/CudaMemoryManager. The facade had exactly one caller
(CudaBackendIntegration, itself only referenced by its own partial declaration)
and its unique surface (GetUsageAnalytics, PerformCleanupAsync) was a
delegation to an internal CudaAsyncMemoryManagerAdapter plus a "simplified"
no-op MemoryPool.

Changes:
- Delete src/Backends/DotCompute.Backends.CUDA/Integration/Components/CudaMemoryManager.cs
- Move supporting public types (MemoryUsageAnalytics, MemoryCleanupSummary,
  MemoryManagementPolicies, MemoryPoolingPolicy, MemoryTrackingPolicy,
  MemoryTransferDirection) into Memory/Models/MemoryAnalyticsTypes.cs to keep
  them available for external consumers.
- Add Memory/CudaMemoryManager.Analytics.cs with GetUsageAnalytics() and
  PerformCleanupAsync() methods on the canonical manager.
- Retarget CudaBackendIntegration to the canonical Memory.CudaMemoryManager.
- Refactor the canonical CudaMemoryManager's Dispose from `new` hiding to the
  standard protected virtual Dispose(bool) pattern so CA2213 recognizes field
  ownership correctly.

Net: -1218 LOC. Single public type named CudaMemoryManager under
DotCompute.Backends.CUDA.Memory. Build 0E/0W, solution-wide.

Part of v1.0.0 GA deduplication pass.
